# Conflicts:
#	client/src/main/java/client/scenes/EventOverviewCtrl.java

+++ client/src/main/java/client/scenes/MainCtrl.java
@@ -17,6 +17,7 @@ package client.scenes;
 
 import client.services.I18N;
 import client.services.NotificationHelper;
+import client.services.NotificationService;
 import client.utils.Config;
 import client.utils.User;
 import commons.Event;
@@ -71,6 +72,7 @@ public class MainCtrl {
 
     private AdminEventsCtrl adminEventsCtrl;
     private Scene adminEvents;
+    private NotificationService notificationService;
 
     public void initialize(Stage primaryStage, Pair<FirstTimeCtrl, Parent> firstTime,
                            Pair<EventOverviewCtrl, Parent> eventOverview,
@@ -157,6 +159,7 @@ public class MainCtrl {
                 }
             }
         });
+        this.notificationService = new NotificationHelper();
     }
 
     public void showInviteView(Event event){
@@ -292,6 +295,9 @@ public class MainCtrl {
     }
 
     public void deleteAllData(){
+        if (!notificationService.showConfirmation("Delete data", "Are you sure you want to delete your data? This action cannot be undone.")){
+            return;
+        }
         Config.deleteUserConfigFile();
         this.chooseFirstPage(false);
     }
